requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
gunicorn==23.0.0
whitenoise==6.11.0
Pillow==12.0.0
xhtml2pdf==0.2.17
sentry-sdk==2.43.0
psycopg2-binary==2.9.11